% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_debug.R
\name{.add_debug}
\alias{.add_debug}
\title{Debug selected functions}
\usage{
.add_debug(fn_name = NULL, debug_live, env)
}
\arguments{
\item{fn_name}{character vector. Names of functions.}

\item{debug_live}{\code{logical} or \code{character}.
\itemize{
  \item{If \code{TRUE}, then all functions with name in \code{fn_name} in environment
  \code{env} are debugged once.
  If \code{FALSE, then none are}}
  \item{If \code{character} and contains \code{'all'}, then all functions with name in \code{fn_name} are debugged once.
  If \code{character} but not including \code{'all'}, then all functions that have names for which the beginning matches
  at least partially any element in \code{debug_live} are debugged once.}
}}

\item{env}{environment. Environment in which functions are found.}
}
\description{
Applies debugonce to specified functions.
}
