% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_dir_proj.r
\name{.create_dir_proj}
\alias{.create_dir_proj}
\title{Create a project directory}
\usage{
.create_dir_proj(dir_base = here::here(), iter = NULL, max_len = NULL)
}
\arguments{
\item{dir_base}{character.
Path to "base" directory, to which
additional sub-directories are appended.
Default is \code{here::here()}.}

\item{iter}{tibble of one row or list.
Each element in iter is used to create
a sub-directory, appended in order from left to
right (i.e. the left-most column of iter creates
the first appended sub-directory).
See \code{?pipeline:::.create_text_ref} for details.}

\item{max_len}{numeric.
Maximum length of the directory.
If \code{NULL}, then is is set to 256
if the operating system is Windows and
1e3 otherwise.
Default is \code{NULL}.}
}
\value{
Invisible returns the project directory.
}
\description{
Create a project directory
}
\examples{
.create_dir_proj(
  dir_base = tempdir(),
  iter = tibble::tibble(
    param1 = c("a"),
    param2 = list(list(fn1 = function(x) x)),
    param3 = c("a" = "a")
  )
)

 iter <- structure(
   list(
     var_exp = "Progressor",
     var_exp_spline = list(
       `tc~soma~C9` = list(fn = "ns", params = list(df = 2))
     ),
     var_int = list(
       c("Progressor", "C9")
     ),
     wins = "wins_n",
     var_conf = list(
       "none"
     ),
     ttb_max = NA_real_,
     var_dep = "risk6"
   ),
   row.names = c(
     NA,
     -1L
   ),
   class = c("tbl_df", "tbl", "data.frame")
 )
.create_dir_proj(dir_base = tempdir(), iter = iter)
}
