% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_out_tbl.r
\name{.get_out_tbl}
\alias{.get_out_tbl}
\title{Collate model output file objects and model parameters}
\usage{
.get_out_tbl(dir_proj, dir_sub = NULL, iter = NULL)
}
\arguments{
\item{dir_proj}{character.
Path to directory where individual
model results are kept.}

\item{dir_sub}{character vector.
Names of directories inside \code{dir_proj}
whose contents are to be included as
relative paths (relative to \code{dir_proj})
inside a column \code{"fn"}.
If \code{NULL}, then all directories
inside \code{dir_proj} are included.
Default is \code{NULL}.}

\item{iter}{dataframe with one row (or,
alternatively, a list that is coerced to
a dataframe (tibble) with one row).
Specifies arguments to specific
model run.}
}
\value{
A tibble with one row, containing the following columns:
path_proj (path to project, relative to package directory,
           i.e. excluding the package name),
fn (a column where each element is a list containing all the relative paths
    to files in specified sub-directories),
and all the columns from \code{iter}, specifying the model setup.
}
\description{
Collate model output file objects and model parameters
}
\details{
Assumes that it is being run inside the R project/package
that created the model.
}
