% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replace_null_fn.R
\name{.replace_null_fns}
\alias{.replace_null_fns}
\title{Replace NULL functions}
\usage{
.replace_null_fns(env, expected_params)
}
\arguments{
\item{env}{environment. Environment in which the functions are to be found and replaced.}

\item{expected_params}{named list. Names are names of functions in \code{env} and values
are character vectors specifying parameters for functions.}
}
\description{
Replaces functions that are NULL or missing in a specified environment with functions whose body is \code{NULL; NULL} and
have empty arguments as specified.
}
\examples{
env_curr <- environment()
fn1 <- function(x) x; fn2 <- function(x = 2) y; fn3 <- NULL
expected_params <- list(fn1 = 'x', fn2 = 'y', fn3 = 'x')
analysispipeline:::.replace_null_fns(env = env_curr, expected_params = expected_params)
fn3 # replaced to have a function with arguments as specified by expected_params
fn2; fn1 # no change in either
}
