% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_fns.R
\name{.validate_fn}
\alias{.validate_fn}
\title{Validate parameters for individual functions}
\usage{
.validate_fn(fn_name, expected_params = NULL, env)
}
\arguments{
\item{fn_name}{character. Name of single function.}

\item{expected_params}{list.
Names are names of functions to be validated
and values are expected parameters.
If \code{NULL}, then a default list is used.
Run \code{body(analysispipeline:::.validate_fns)} to see default.
Default is \code{NULL}.}

\item{env}{environment.
Environment in which the functions are found and to be saved.}
}
\description{
Validate parameters for individual functions
}
