% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_fns.R
\name{.validate_fns}
\alias{.validate_fns}
\title{Check that functions have correct parameters}
\usage{
.validate_fns(env, expected_params = NULL)
}
\arguments{
\item{env}{environment. Environment in which the functions are found and to be saved.}

\item{expected_params}{list. Names are names of functions to be validated and values are expected parameters.
If \code{NULL}, then a default list is used. Run \code{body(analysispipeline:::.validate_fns)} to see default.
Default is \code{NULL}.}
}
\description{
Each function supplied to \code{run_analysis_pipeline} must have certain parameters.
They alaways include \code{params_dots}.
}
\examples{
env_curr <- environment()
fn1 <- function(x) x; fn2 <- function(x) y
expected_params <- list(fn1 = 'x', fn2 = 'y')
# error
testthat::expect_error(analysispipeline:::.validate_fns(env = env_curr,
                                 expected_params = expected_params))

# no error given if fn is NULL
fn2 <- NULL
analysispipeline:::.validate_fns(env = env_curr,
                                 expected_params = expected_params)
}
