% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_expected_params.R
\name{get_expected_params}
\alias{get_expected_params}
\title{Print the expected parameters for each function}
\usage{
get_expected_params()
}
\description{
Print the names of each function
along with the respective expected parameters.
}
