% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manual_marker_sets-classes.R
\name{get_markers}
\alias{get_markers}
\alias{print.marker_set}
\title{Get marker set from \code{manual_datasets()}}
\usage{
get_markers(name)

\method{print}{marker_set}(x, ...)
}
\arguments{
\item{name}{The name of the marker set to extract. \code{name} usually is the
surname of the first author followed by the years of the published article
connected with "_".}

\item{x}{An \code{marker_set} object}

\item{...}{Not used currently}
}
\value{
A \code{marker_set} object.
}
\description{
The package collected some marker sets from research articles in
\code{manual_datasets()}, this function just provide a method to extract the
specific marker set from the datasets
}
