% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/persons_wikidata.R
\name{init_person_wikidata_table}
\alias{init_person_wikidata_table}
\title{Initialize person_wikidata table}
\usage{
init_person_wikidata_table(con)
}
\arguments{
\item{con}{DuckDB connection}
}
\value{
TRUE invisibly
}
\description{
Initialize person_wikidata table
}
