% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manual_marker_sets-classes.R
\name{manual_datasets}
\alias{manual_datasets}
\alias{print.manual_cell_marker_datasets}
\title{cell markers datasets collected from research articles}
\usage{
manual_datasets()

\method{print}{manual_cell_marker_datasets}(x, ...)
}
\arguments{
\item{x}{An \code{manual_cell_marker_datasets} object.}

\item{...}{Not used currently}
}
\value{
An \code{manual_cell_marker_datasets} object.
}
\description{
We manullay collected some marker sets from research articles,
\code{manual_datasets} function just returns the database with a class of
\code{manual_cell_marker_datasets} which has a specific \code{print} method.
}
