% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/open_folder.R
\name{open_folder}
\alias{open_folder}
\alias{open_dir}
\title{Open a dialogue box at selected folder}
\usage{
open_folder(path = rstudioapi::getActiveProject, ...)

open_dir(path = rstudioapi::getActiveProject, ...)
}
\arguments{
\item{path}{character. The initial working directory, from which the file dialog should begin browsing. Defaults to the current RStudio project directory.}

\item{...}{
  Arguments passed on to \code{\link[rstudioapi:file-dialogs]{rstudioapi::selectFile}}
  \describe{
    \item{\code{caption}}{The window title.}
    \item{\code{label}}{The label to use for the 'Accept' / 'OK' button.}
    \item{\code{filter}}{A glob filter, to be used when attempting to open a file with a
particular extension. For example, to scope the dialog to \R files, one could use
\code{R Files (*.R)} here.}
    \item{\code{existing}}{Boolean; should the file dialog limit itself to existing
files on the filesystem, or allow the user to select the path to a new file?}
  }}
}
\value{
\code{invisible(<selected_file>/NULL)} (\code{NULL} if no file were
  selected).
}
\description{
Useful for easily opening a window in
which you can see all the files in a folder and open them using
\code{Open with...}.
}
\details{
Only works in RStudio. Wrapper around `rstudioapi::selectFile` that simply
passes all its arguments to it but avoids printing the path to a file
if a file were selcted.
}
\examples{
open_folder()
open_dir()
open_folder(.libPaths()[1])
}
