% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edirc.R
\name{parse_all_institutions}
\alias{parse_all_institutions}
\title{Parse the EDIRC archive into an institutions tibble}
\usage{
parse_all_institutions(repec_folder = NULL, script_path = NULL)
}
\arguments{
\item{repec_folder}{Path to the root RePEc folder. Defaults to config$repec_folder}

\item{script_path}{Path to parse_redif_dir.pl. Defaults to the bundled script}
}
\value{
Institutions tibble (invisibly)
}
\description{
Parse the EDIRC archive into an institutions tibble
}
