% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline-package.R
\docType{package}
\name{pipeline-package}
\alias{pipeline-package}
\alias{_PACKAGE}
\title{pipeline: R package to create pipeline}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Yun Peng \email{yunyunpp96@outlook.com}

}
\keyword{internal}
