% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edirc.R
\name{post_process_institution_entry}
\alias{post_process_institution_entry}
\title{Post-process a parsed ReDIF-Institution entry}
\usage{
post_process_institution_entry(entry)
}
\arguments{
\item{entry}{Parsed entry list from parse_redif_dir.pl}
}
\value{
One-row tibble or NULL if invalid
}
\description{
Post-process a parsed ReDIF-Institution entry
}
