% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_in_dir.R
\name{print_plots_in_dir}
\alias{print_plots_in_dir}
\title{Output code to print plots in a directory in a Knitr document}
\usage{
print_plots_in_dir(dir_plot, level = 3, format = c("png", "pdf"))
}
\description{
Output code to print plots in a directory in a Knitr document
}
