% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_in_dir.R
\name{print_tbl_in_dir}
\alias{print_tbl_in_dir}
\title{Output code to print tables in a directory in a Knitr document}
\usage{
print_tbl_in_dir(dir_tbl, level = 3)
}
\description{
Output code to print tables in a directory in a Knitr document
}
