% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_in_dir.R, R/save_objects.R
\name{return_rds}
\alias{return_rds}
\title{Return RDS objects in a folder}
\usage{
return_rds(dir_obj)

return_rds(dir_obj)
}
\value{
A named list, where the names are the names of the files and
the elements are the RDS objects. If there are no RDS files in folder,
then an empty list is returned.

A named list, where the names are the names of the files and
the elements are the RDS objects. If there are no RDS files in folder,
then an empty list is returned.
}
\description{
Return RDS objects in a folder

Return RDS objects in a folder
}
