% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_analysis_pipeline.R
\name{run_analysis_pipeline}
\alias{run_analysis_pipeline}
\title{Run GAMLSS pipeline}
\usage{
run_analysis_pipeline(
  dir_proj,
  dir_proj_empty = FALSE,
  data_raw,
  preprocess_fn = NULL,
  plot_exp_fn = NULL,
  fit_fn = NULL,
  get_fit_stats_fn = NULL,
  plot_fit_fn = NULL,
  plot_val_fn = NULL,
  debug_live = FALSE,
  ...
)
}
\arguments{
\item{dir_proj}{\code{character} or \code{function}.
\itemize{
  \item{If \code{character}, then this folder is created if it doesn't exist.}
  \item{If a \code{function}, then it should be a function that
  has a single argument that expects the named list produced by \code{rlang::list2(...)}.
  This function should simply return. the project directory as a character.
  This folder is created if it doesn't exist.}
}}

\item{dir_proj_empty}{\code{logical}. If \code{TRUE}, then the project directory
is emptied before any results are saved. Default is \code{FALSE}.
TODO: Wrap each of the *_fn functions so that whatever they return is saved to the
correct folder.
TODO: Enable debugging via a parameter even sub-functions of the main functions.
TODO: Ensure that functions are not run in global environment but run in own environment (or a clean environment, as in knitr);
if simply ensuring that they're not run in global environment, then add the function
to save object to global environment if need be, and then delete it upon function
exit.}

\item{data_raw}{dataframe. Contains raw data to be processed by \code{preprocess_fn} to create \code{data_mod}.}

\item{preprocess_fn, plot_exp_fn, fit_fn, get_fit_stats_fn, plot_fit_fn, plot_val_fn}{functions or \code{NULL}. Functions
run in the order specified, passing data_raw, data_mod and params_dots from one to the other. Some functions
expect extra output.}

\item{debug_live}{character vector or \code{FALSE}. Functions provided to \code{run_analysis_pipeline}
that at least partially match characters in \code{debug_live} at the start are run in \code{debugonce}.}

\item{...}{Named arguments passed onto all functions (preprocess_fn to plot_val_fn).}
}
\description{
Takes arguments that specify the running of the key steps in running a GAMLSS pipeline:
\itemize{
  \item{project directory creation,}
  \item{data pre-processing,}
  \item{creating exploratory plots,}
  \item{fitting the model,}
  \item{creating validation plots, and}
  \item{creating results plots.}
}
The purpose of this is to make it easier to do all the steps in a reproducible way that
saves the results automatically.
}
