% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_objects.R
\name{save_objects}
\alias{save_objects}
\title{Save objects to specified directory}
\usage{
save_objects(
  ...,
  obj_list = NULL,
  dir_proj,
  dir_sub = NULL,
  empty = TRUE,
  silent = FALSE,
  gg_device = "png",
  embed = TRUE,
  height = 10,
  width = 12
)
}
\arguments{
\item{...}{Objects to be saved. Preferrably specify name-value pairings,
otherwise default values will be given. Objects are saved with the name appended by ".rds".}

\item{obj_list}{list. List where the names (if supplied) are the
names to save the corresponding values in the list as. If names
are not supplied, then they will be created. Note that if you want to save ggplot2 plots as "PNG" or "PDF" images with
custom names, then you need to pass them to the function as a named list to the \code{obj_list} parameter, e.g.
\code{obj_list = list("Exploratory plot" = p_exp)} will save the plot \code{p_exp} with name "Exploratory plot.png" (or .pdf, if
that's chosen below.)}

\item{dir_proj}{character.}

\item{dir_sub}{character vector or character list. Each element appends a new directory
to dir_proj, in the order they're given. If not provided, then \code{dir_proj} is the final
directory.}

\item{empty}{logical. If \code{TRUE}, then the
directory to be saved to is attempted to be deleted before
saving any new files to it. This attempt
fails on Windows 10, at least, if any files
from these directories are open, or even if just
a Windows Explorer object is open to that directory or a
a sub-directory. Default is \code{TRUE}.}

\item{silent}{logical. If \code{FALSE}, then warnings and messages
are suppressed. Default is \code{TRUE}.}

\item{gg_device}{'rds', 'png' or 'pdf. Specifies device to
save objects of class \code{'gg'}. They are saved using \code{cowplot::ggsave2} function. If
gg_device does not contain any of rds, png or pdf, then saved as rds. Default is \code{'png'}.}

\item{width, height}{numeric. Height and width to pass to \code{cowplot::ggsave2}. Default is
12 and 10.}
}
\value{
\code{invisible{TRUE}}.
}
\description{
Saves objects specified using dots and/or a list.
Ensures that all names are unique (appending with "-uni_<ind>" for
duplicated names). Ensures that all objects have names (appending with "-obj_<ind>")
for objects without names.
}
\details{
Files that already exist before they are saved are deleted. This is to ensure
that on Windows the "Date Modified" column in Windows Explorer updates.
}
\examples{
# standard use
dir_proj <- tempdir()
x <- "a"; y <- "b"
save_objects(x = x, y = y, dir_proj = dir_proj, dir_sub = "test")
for(obj in c('x', 'y')) print(file.exists(file.path(dir_proj, "test",
                                                    paste0(obj, ".rds"))))
readRDS(file.path(dir_proj, "test", "x.rds"))
readRDS(file.path(dir_proj, "test", "y.rds"))

# ensure all objects have names
# warning normally printed
suppressWarnings(save_objects('a', obj_list = list('c', 100, z = 4),
                              dir_proj = dir_proj, dir_sub = "test"))
readRDS(file.path(dir_proj, "test", "obj_1.rds"))
readRDS(file.path(dir_proj, "test", "obj_2.rds"))
readRDS(file.path(dir_proj, "test", "obj_3.rds"))
# objects with names given do not have names overwritten
readRDS(file.path(dir_proj, "test", "z.rds"))

# ensure that objects have unique names
# warning printed
save_objects(x = 'a', x = 'b', dir_proj = dir_proj, dir_sub = "test")
readRDS(file.path(dir_proj, "test", "x-uni_1.rds"))
readRDS(file.path(dir_proj, "test", "x-uni_2.rds"))
}
