% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scp-package.R
\docType{package}
\name{scp-package}
\alias{scp-package}
\alias{_PACKAGE}
\title{scp: Single Cell Pipeline}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Yun Peng \email{yunyunpp96@outlook.com}

}
\keyword{internal}
