% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-scp.R
\name{scp}
\alias{scp}
\title{Single Cell Pipeline}
\usage{
scp(data = list(), ...)
}
\arguments{
\item{data}{A list of object used in the pipeline.}

\item{...}{<\code{\link[rlang:dyn-dots]{dynamic-dots}}>, all items must be a \code{step}
object. Can also provide as a list of steps directly.}
}
\value{
A \code{Pipeline} object.
}
\description{
Single Cell Pipeline
}
