% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-step.R
\name{step}
\alias{step}
\alias{create_step}
\alias{[[.step}
\alias{[[<-.step}
\alias{$.step}
\alias{$<-.step}
\alias{[.step}
\alias{[<-.step}
\title{Build a \code{step} object}
\usage{
step(id, call, deps = NULL, finished = FALSE, return = TRUE, seed = FALSE, ...)

create_step(
  id,
  call,
  deps = NULL,
  finished = FALSE,
  return = TRUE,
  seed = FALSE,
  ...
)

\method{[[}{step}(x, i)

\method{[[}{step}(x, i) <- value

\method{$}{step}(x, i)

\method{$}{step}(x, i) <- value

\method{[}{step}(x, i)

\method{[}{step}(x, i) <- value
}
\arguments{
\item{id}{A scalar character indicates the identification of the step. Must
be unique across the \code{step_tree}.}

\item{call}{The call define the command to run. Function \code{step} will use
<\code{\link[rlang:enquo]{rlang::enquo()}}> to defuse this argument.}

\item{deps}{A character vector or \code{NULL} defines the upstream steps to run
before runing this step. \code{NULL} means no dependencies.}

\item{finished}{A scalar \code{logical} indicates whether this step has been
evaluated.}

\item{return}{A scalar \code{logical} indicates whether to keep the returned
value. if \code{FALSE}, the result evaluated from \code{call} won't be kept.}

\item{seed}{A scalar \code{logical} or \code{numeric}. a \code{logical} value indicates
whether to set seed when evaluated the expression in "call". if \code{TRUE}, the
call is evaluated with a seed (based on the hash of the call object).
Otherwise, the call is evaluated without seed. if \code{numeric}, seed will be set
by \code{set.seed(as.integer(seed))}.}

\item{...}{<\code{\link[rlang:dyn-dots]{dynamic-dots}}> Other items to extend \code{step}
object.}

\item{x}{A \code{step} object from which to extract element(s) or in which to
replace element(s).}

\item{i}{The indices specifying elements to extract or replace}

\item{value}{The value to replace with.}
}
\value{
A \code{step} object.
}
\description{
A \code{step} object define the command to run in the pipeline.

Both \code{step} and \code{create_step} do the same thing, except that \code{step} function
also defuse the \code{call} argument as a quosure.
}
