% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_dir.r
\name{clean_dir}
\alias{clean_dir}
\title{Clean up a directory}
\usage{
clean_dir(x, fn, base_name_exception = NULL)
}
\arguments{
\item{x}{character.
Directory to clea.}

\item{fn}{character vector.
Paths (preferrably absolute) to files
that should not be deleted.}

\item{base_name_exception}{character vector.
File names that are not to be deleted,
(even if not specified within fn).
Matched on the basis of regular expression.}
}
\value{
If successful, \code{invisible(TRUE)}.
}
\description{
Remove all unspecified files, or files
not matching a particular character expression, from
a directory.
}
