% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_proj_dir.R
\name{.setup_proj_dir}
\alias{.setup_proj_dir}
\title{Set up project directory}
\usage{
.setup_proj_dir(dir_proj, dir_proj_empty, p_dots)
}
\arguments{
\item{p_dots}{named list. Named list returned by \code{rlang::list2} applied to
the ellipses of \code{run}.}
}
\value{
A \code{character}, that is the absolute path of the project directory. Side effects
are that this project directory is created if it doesn't exist and emptied if \code{proj_dir_empty}
is \code{TRUE}.
}
\description{
Get the project directory, either as supplied (if \code{dir_proj} in \code{run} is
a \code{character}) or as created by the \code{dir_proj} parameter if its argument is a \code{function}
taking its arguments from the dots of \code{run}.

This directory is always created, and emptied if instructed.
}
