% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qsub.R
\name{qsub}
\alias{qsub}
\title{Run expressions on a gridengine system}
\usage{
qsub(..., node = NULL, name = NULL, wd = getwd(), resource_list = character())
}
\arguments{
\item{...}{Expressions to defuse.}

\item{node}{Node to run the expressions.}

\item{name}{A string define the job name.}

\item{wd}{Path define the working directory.}

\item{resource_list}{A named characters define the given resources.}
}
\description{
Run expressions on a gridengine system
}
